refactor: make drawPdf function less opinionated
I think drawPdf should leave the styling and scale of the pdf to the hook caller.  

Setting the canvas height and width is fine for displaying the PDF since the canvas is taking on the height and width of the PDF. The canvas element that uses the hook can be styled in any fashion through class name or style.

The default scale of 1 is acceptable for any screen. Making the caller of the hook responsible for the scale allows for better usability of the hook in many cases.